refactor: isolate summary utils
